+++ src/repl/core.rs
@@ -51,11 +51,6 @@ pub fn repl(session: Session) -> Result<(), Signal> {
 
                         match stack.eval_and_finalize(expr) {
                             Err(Signal::Condition(Cond::Terminate)) => break,
-                            Err(Signal::Return(value, true)) => match value {
-                                Obj::Null => (),
-                                _ => print!("{value}"),
-                            },
-                            Err(Signal::Return(_value, false)) => (),
                             Err(e) => {
                                 print!("{e}");
                                 print!("backtrace:\n{stack}");